Fixed and documented replication example for v10
