--- examples-nodejs-v10/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "hypercore-protocol-nodejs-v10",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "GPL-3.0",
-  "dependencies": {
-    "chalk": "^4.0.0",
-    "hypercore": "^10",
-    "pretty-bytes": "^5.3.0",
-    "random-access-memory": "^6.0.0",
-    "split2": "^3.1.1"
-  }
-}
