+++ examples-nodejs/package.json
@@ -1,20 +1,14 @@
 {
-  "name": "hypercore-protocol-nodejs",
+  "name": "hypercore-protocol-rs-nodejs",
   "version": "1.0.0",
-  "description": "",
+  "description": "hypercore-protocol-rs NodeJS example",
   "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "GPL-3.0",
+  "license": "MIT",
   "dependencies": {
     "chalk": "^4.0.0",
-    "hypercore": "^9.0",
-    "hypercore-protocol": "^8.0",
+    "hypercore": "^10",
     "pretty-bytes": "^5.3.0",
-    "random-access-memory": "^3.1.1",
+    "random-access-memory": "^6.0.0",
     "split2": "^3.1.1"
   }
 }
